+++ squirrel-foundation/src/main/java/org/squirrelframework/foundation/fsm/MutableTimedState.java
@@ -2,9 +2,9 @@ package org.squirrelframework.foundation.fsm;
 
 public interface MutableTimedState<T extends StateMachine<T, S, E, C>, S, E, C> extends MutableState<T, S, E, C> {
     
-    void setTimeInterval(Integer timeInterval);
+    void setTimeInterval(long timeInterval);
     
-    void setInitialDelay(Integer initialDelay);
+    void setInitialDelay(long initialDelay);
     
     void setAutoFireEvent(E event);
